linux: Add support for applying DTOs for add-on cards on PDK2
Add minor patch to Linux kernel to permit building dtbo files from dts
sources, so that the kernel-fitimage bbclass can run make ... foo.dtbo
and obtain a DT overlay.

Build such overlays as separate configurations into the kernel fitImage,
one for each add-on card which could be plugged into the PDK2 carrier
board headers. This permits U-Boot to apply these DT overlays before
Linux is booted and then pass such a DT to Linux, and thus avoids the
need for having DT overlay support in Linux directly, which has various
drawbacks and limitations. Instead, the user can pick which base
configuration DT of the PDK2 and which add-on card DT overlays are to
be applied by passing fitImage configurations to 'bootm' command.

The following example selects the base PDK2 carrier board with an
DH 460-200 SRAM add-on card plugged into header X11:

 => bootm ${loadaddr}#conf@stm32mp157c-dhcom-pdk2.dtb#stm32mp157c-dhcom-pdk2-overlay-460-200-x11.dtbo

Currently, the following add-on options are available:
 - DH 460-200 SRAM board    in header X11
 - DH 531-100 SPI/I2C board in header X21
 - DH 531-200 SPI/I2C board in header X22

Signed-off-by: Marek Vasut <marex@denx.de>