linux: Add DT bindings for interfaces on AV96 LS expansion
Add DT bindings for interfaces on the Avenger96 LS expansion connector,
I2C2, FDCAN1, FDCAN2, ADC and SPI2. The SPI2 is disabled by default as
there is no device connected to it. To handle the SPI2, the best way is
to apply a DTO in U-Boot which binds a device and enables the SPI2.

Signed-off-by: Marek Vasut <marex@denx.de>