linux: Add fake pressure reporting for ft5x06 touchscreen on qt4
The qt4 does not work with multi-touch capacitive touchscreens, hence
it is necessary to patch the kernel such that it reports also various
legacy events which the qt4 needs. This adds such a kernel patch only
for the PDK2 and only in case the qt4 layer is activated, to avoid
tainting the kernel otherwise.

Signed-off-by: Marek Vasut <marex@denx.de>