linux: Add GPIO keys and LEDs on PDK2
Add DT bindings for GPIO keys and LEDs on PDK2. The LED5 GPIO-E has a
conflict with touchscreen IRQ on DH 560-200 display unit, so the DTO
for that display unit disables the LED5. TA1 button has an IRQ line
conflict with ethernet, so TA1 button is polled. TA2, TA3, TA4 buttons
serve as wakeup sources.

Signed-off-by: Marek Vasut <marex@denx.de>