linux: Invert pixel clock polarity on 505-200 RGB-to-LVDS adapter
The OnSemi FIN3385 Parallel-to-LVDS encoder on DH 505-200 adapter board
is strapped to sample pixel data on rising edge of the input clock using
R57 resistor, which is pulling FIN3385 RF input line high.

Add patch to permit configuration of pixel data sampling edge in the
LVDS encoder driver, repair LTDC driver to respect the DRM bus flags
provided by either the display or encoder, and adapt the 505-200 DTO
to indicate the LVDS encoder is strapped such that it samples pixel
data on rising edge of the input clock.

Signed-off-by: Marek Vasut <marex@denx.de>