linux: Repair RSI9116 SDIO WiFi driver association with IWL WiFi
In case RSI9116 SDIO WiFi operates in STA mode against Intel 9260 in AP mode,
the association fails. The former is using wpa_supplicant during association,
the later is set up using hostapd.

The problem is that the TX EAPOL data descriptor RSI_DESC_REQUIRE_CFM_TO_HOST
flag and extended descriptor EAPOL4_CONFIRM frame type are not set in case the
AP is iwlwifi, because in that case the TX EAPOL packet is 2 bytes shorter.

The downstream vendor driver has this change in place already, however there
is no explanation for it, neither is there any commit history from which such
explanation could be obtained.

Signed-off-by: Marek Vasut <marex@denx.de>