linux: Enable voluntary preemption and increase HZ to 250
The kernel is currently configured as non-preemptible on all DHSOM, this
means the system can use maximum of CPU time for computation, while the
scheduling latencies are higher, which is useful on server systems. This
patch sets the system as voluntary-preemptible, which means the kernel
has more preemption points and the system can react with lower latency,
which is a better fit for embedded systems. This patch also increases HZ
from 100 to 250, which also improves latency without spending too much
time on scheduling.

Signed-off-by: Marek Vasut <marex@denx.de>