linux: Move RSI WiFi interrupt handling to RX thread
The interrupt handling of the RS911x is particularly heavy. For each RX
packet, the card does three SDIO transactions, one to read interrupt
status register, one to RX buffer length, one to read the RX packet(s).
This translates to ~330 uS per one cycle of interrupt handler. In case
there is more incoming traffic, this will be more. This leads to a very
long SDIO interrupt, which in turns leads to a poor WiFi performance.
Fix this by moving the entire interrupt handling for the RSI SDIO WiFi
into RX thread instead.

Signed-off-by: Marek Vasut <marex@denx.de>