linux: Unify kernel defconfigs
Merge all the kernel defconfigs into a single common DHSOM defconfig to make
the defconfig situation easier to handle. This does lead to a couple of extra
kernel modules being installed on all the systems, however for evaluation
purposes this causes no issues and a final system with DHSOM is expected to
use a custom-tailored optimal defconfig.

Signed-off-by: Marek Vasut <marex@denx.de>