u-boot: Add support for applying custom DT changes before boot
This patch to the U-Boot boot script permits user to set an U-Boot script
named "loaddtoscustom", which performs modifications to a DT just before
the Linux kernel is started. Such modifications are generally implemented
using the U-Boot "fdt" command and range from simple DT node editing all
the way to applying custom DT overlays loaded from EEPROMs, network, etc.

In case the "loaddtoscustom" script is not set, the boot process is the
same as it used to be before, DT and DTOs are applied through the bootm
command directly from the fitImage.

In case the "loaddtoscustom" script is set, the boot process is altered.

In case the "loaddtos" variable is not empty, the base DT and DTOs are
extracted from the fitImage, the "fdt" command is configured to point to
the base DT, and the DTOs are applied in order to the base DT using the
"fdt apply" command. The address of the DT with the select DTOs applied
is stored in the "fdtaddr" variable.

In case the "loaddtos" variable is empty, none of the above setup is
performed, and it is up to the user to set up the "fdt" command in the
"loaddtoscustom" script, e.g. by using "fdt addr" to point to a custom
DT.

In either case, the "loaddtoscustom" script is executed last and permits
the user to perform modifications to the resulting DT. The kernel is then
booted using "bootm ${bootm_args}", the "bootm_args" is by default set to
"${loadaddr} - ${fdtaddr}", however it is possible to override this too.

Example:

=> setenv loaddtoscustom 'fdt set / model hello'
=> boot
...
[    0.000000] Booting Linux on physical CPU 0x0
...
[    0.000000] OF: fdt: Machine model: hello

Signed-off-by: Marek Vasut <marex@denx.de>