linux: Enable ramdisk support
Enable support for ramdisk and loopback devices in the Linux kernel.
This is required for booting system with a separate initrd that is
passed to the kernel as a memory location, e.g. an initrd embedded
in fitImage and passed to the kernel by U-Boot. This is not related
to initramfs embedded in the Linux kernel.

Signed-off-by: Marek Vasut <marex@denx.de>