u-boot: Handle empty loaddtos if loaddtoscustom is set
The $loaddtoscustom U-Boot environment variable may contain optional
script, which is used to modify a DT just before the DT is passed to
the Linux kernel and the kernel is booted. Such a DT which is input
into the $loaddtoscustom script is assembled according to the content
of $loaddtos variable, and is composed of base DT, with optional DT
overlays applied on top. Hence, currently the $loaddtoscustom script
depends on $loaddtos variable to contain at least one base DT entry;
if $loaddtos is empty, $loaddtoscustom script cannot work.

This is inconvenient and inobvious. This patch adds handling of the
case where $loaddtos is empty and $loaddtoscustom is set by setting
default value to $loaddtos and then handling it as if $loaddtos was
set with only a base DT entry. The default value is extracted from
kernel fitImage default configuration node. The script also notifies
user of this and the user can override the default value by setting
the $loaddtos variable as before.

Note that in case $loaddtoscustom is not set, $loaddtos variable may
or may not be set. Such a case is not affected by this modification
or this issue.

Note that due to change in indentation, this patch is better viewed
using 'git show -w'.

Signed-off-by: Marek Vasut <marex@denx.de>