linux: Assign RSI beacon rate settings to the correct rate_info descriptor field
The RSI_RATE_x bits must be assigned to struct rsi_data_desc rate_info
field. The rest of the driver does it correctly, except this one place,
so fix it. This is also aligned with the RSI downstream vendor driver.
Without this patch, an AP operating at 5 GHz does not transmit any
beacons at all, this patch fixes that.

Signed-off-by: Marek Vasut <marex@denx.de>