Correct IMAGE_INSTALL variable so not to fail when apt-get installs packages.
During do_rootfs apt-get install command complains about not finding the package 'virtual'. The apt-get command is constructed by directly expanding IMAGE_INSTALL into the target packages for apt to install, so it gets 'virtual/bootloader' as argument, hence its failure. RPM seems to work with that with no problems.

By this modification it's even easier to override the desired bootloader in a user-added layer.