machine: Prefer libdrm-mainline and wayland-protocols-mainline for dunfell
Since meta-mainline-common mesa 22.0.y update, the mesa requires
up-to-date version of libdrm and wayland-protocols. Those are now
available as backports in meta-mainline-common as well, but they
are not preferred by default. Set the default preferrence to pull
in the up-to-date versions in case oe-core release is dunfell. The
kirkstone release contains up to date versions already.

Signed-off-by: Marek Vasut <marex@denx.de>