cypress-firmware: Install board specific CLM and firmware blob symlinks
Install board specific CLM and firmware blob symlinks so that the
brcmfmac driver can pick them up immediatelly instead of waiting
for the fallback to start and pick generic blob. The fallback takes
about a minute to start, which makes the WiFi MAC interface appear
about a minute after boot. Improve this.

Signed-off-by: Marek Vasut <marex@denx.de>