README: Document testbench board support
Document support for DHCOR Testbench board, which is a testbench for
testing of DHCOR SoM during manufacturing. This is effectively a trimmed
down version of AV96 board with CSI-2 bridge, HDMI bridge, WiFi, Audio
and LEDs removed and used as GPIOs instead. Furthermore, the PMIC Buck3
is always configured from PMIC NVM to cater for both 1V8 and 3V3 SoM
variant.

Signed-off-by: Marek Vasut <marex@denx.de>