perf: Enable jevents PACKAGECONFIG
The perf PMU events and metrics are generated at build time using jevents.py
script, this is only triggered if jevents PACKAGECONFIG is present, which is
not the case by default.

The script processes json files in the kernel tree like these:
tools/perf/pmu-events/arch/arm64/freescale/imx8mp/sys/metrics.json
which then augment perf processing capabilities with metrics, which
allow for better filtering and interpretation of the raw hardware
counters.

Enable jevents PACKAGECONFIG so the metrics are generated and included
in the perf build.

Signed-off-by: Marek Vasut <marex@nabladev.com>